pyproject.toml
[tool.poetry]
name = "webscraping-product-bot"
version = "0.1.0"
description = ""
authors = ["lksprado <lks-prado@live.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
pandas = "^2.2.3"
psycopg2-binary = "^2.9.10"
lxml = "^5.3.0"
requests-html = "^0.10.0"
lxml-html-clean = "^0.4.1"
selenium = "^4.27.1"
undetected-chromedriver = "^3.5.5"
setuptools = "^75.6.0"
python-telegram-bot = "^21.10"
playwright = "^1.49.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
